+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Release channels have their own copy of this changelog:
   * `central-scheduler` as default option for `--block-production-method` (#34891)
   * `solana-rpc-client-api`: `RpcFilterError` depends on `base64` version 0.22, so users may need to upgrade to `base64` version 0.22
   * Changed default value for `--health-check-slot-distance` from 150 to 128
+  * CLI: Can specify `--with-compute-unit-price` and `--max-sign-attempts` during program deployment
 
 ## [1.18.0]
 * Changes
@@ -39,6 +40,7 @@ Release channels have their own copy of this changelog:
     double the size. Program accounts must be extended with `solana program extend`
     before an upgrade if they need to accommodate larger programs.
   * Interface for `gossip_service::get_client()` has changed. `gossip_service::get_multi_client()` has been removed.
+  * CLI: Can specify `--with-compute-unit-price` and `--max-sign-attempts` during program deployment
 * Upgrade Notes
   * `solana-program` and `solana-sdk` default to support for Borsh v1, with
 limited backward compatibility for v0.10 and v0.9. Please upgrade to Borsh v1.